+++ dialer.go
@@ -40,6 +40,7 @@ import (
 	"cloud.google.com/go/cloudsqlconn/instance"
 	"cloud.google.com/go/cloudsqlconn/internal/cloudsql"
 	"cloud.google.com/go/cloudsqlconn/internal/mdx"
+	"cloud.google.com/go/cloudsqlconn/internal/sqldataclient"
 	"cloud.google.com/go/cloudsqlconn/internal/trace"
 	"github.com/google/uuid"
 	"golang.org/x/net/proxy"
@@ -48,6 +49,8 @@ import (
 	tel "cloud.google.com/go/cloudsqlconn/internal/tel"
 	monitoring "cloud.google.com/go/monitoring/apiv3/v2"
 	sqladmin "google.golang.org/api/sqladmin/v1beta4"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 )
 
 const (
@@ -156,10 +159,11 @@ func (c *dialerConfig) getClientUniverseDomain() string {
 //
 // Use NewDialer to initialize a Dialer.
 type Dialer struct {
-	lock           sync.RWMutex
-	cache          map[cacheKey]*monitoredCache
-	keyGenerator   *keyGenerator
-	refreshTimeout time.Duration
+	lock               sync.RWMutex
+	cache              map[cacheKey]*monitoredCache
+	sqlDataConnAllowed map[cacheKey]bool
+	keyGenerator       *keyGenerator
+	refreshTimeout     time.Duration
 	// closed reports if the dialer has been closed.
 	closed chan struct{}
 
@@ -192,11 +196,17 @@ type Dialer struct {
 
 	// resolver converts instance names into DNS names.
 	resolver       instance.ConnectionNameResolver
+	dnsResolver    cloudsql.NetResolver
 	failoverPeriod time.Duration
 
 	// metadataExchangeDisabled true when the dialer should never
 	// send MDX mdx requests.
 	metadataExchangeDisabled bool
+
+	sqlDataDialer sqldataclient.Dialer
+
+	tokenProvider   auth.TokenProvider
+	sqlDataEndpoint string
 
 	// applicationName is the name of the application using the dialer.
 	applicationName string
@@ -227,12 +237,15 @@ func (nullLogger) Debugf(_ context.Context, _ string, _ ...interface{}) {}
 // RSA keypair is generated will be faster.
 func NewDialer(ctx context.Context, opts ...Option) (*Dialer, error) {
 	cfg := &dialerConfig{
-		refreshTimeout:  cloudsql.RefreshTimeout,
-		dialFunc:        proxy.Dial,
-		logger:          nullLogger{},
-		useragents:      []string{userAgent},
-		failoverPeriod:  cloudsql.FailoverPeriod,
-		applicationName: "unknown",
+		refreshTimeout:       cloudsql.RefreshTimeout,
+		dialFunc:             proxy.Dial,
+		logger:               nullLogger{},
+		useragents:           []string{userAgent},
+		failoverPeriod:       cloudsql.FailoverPeriod,
+		dnsResolver:          net.DefaultResolver,
+		sqlDataEndpoint:      "sqladmin.googleapis.com",
+		sqlDataStreamTimeout: 2 * time.Hour,
+		applicationName:      "unknown",
 	}
 	for _, opt := range opts {
 		opt(cfg)
@@ -313,9 +326,9 @@ func NewDialer(ctx context.Context, opts ...Option) (*Dialer, error) {
 	}
 
 	dc := dialConfig{
-		ipType:       cloudsql.PublicIP,
-		tcpKeepAlive: defaultTCPKeepAlive,
-		useIAMAuthN:  cfg.useIAMAuthN,
+		connectionType: cloudsql.PublicIP,
+		tcpKeepAlive:   defaultTCPKeepAlive,
+		useIAMAuthN:    cfg.useIAMAuthN,
 	}
 	for _, opt := range cfg.dialOpts {
 		opt(&dc)
@@ -336,8 +349,21 @@ func NewDialer(ctx context.Context, opts ...Option) (*Dialer, error) {
 		r = cfg.resolver
 	}
 
+	tp := cfg.apiTokenProvider
+	if tp == nil && cfg.authCredentials != nil {
+		tp = cfg.authCredentials.TokenProvider
+	}
+
+	var sqlDataDialer sqldataclient.Dialer
+	if cfg.sqlDataDialer != nil {
+		sqlDataDialer = cfg.sqlDataDialer
+	} else {
+		sqlDataDialer = sqldataclient.NewGrpcDialer(cfg.sqlDataEndpoint, tp, cfg.quotaProject, cfg.logger, false, cfg.sqlDataStreamTimeout, strings.Join(cfg.useragents, " "))
+	}
+
 	d := &Dialer{
 		closed:                   make(chan struct{}),
+		sqlDataConnAllowed:       make(map[cacheKey]bool),
 		cache:                    make(map[cacheKey]*monitoredCache),
 		lazyRefresh:              cfg.lazyRefresh,
 		disableBuiltInMetrics:    cfg.disableBuiltInMetrics,
@@ -351,9 +377,13 @@ func NewDialer(ctx context.Context, opts ...Option) (*Dialer, error) {
 		iamTokenProvider:         cfg.iamLoginTokenProvider,
 		metricRecorders:          map[instance.ConnName]tel.MetricRecorder{},
 		dialFunc:                 cfg.dialFunc,
+		dnsResolver:              cfg.dnsResolver,
 		resolver:                 r,
 		failoverPeriod:           cfg.failoverPeriod,
 		metadataExchangeDisabled: cfg.metadataExchangeDisabled,
+		tokenProvider:            tp,
+		sqlDataEndpoint:          cfg.sqlDataEndpoint,
+		sqlDataDialer:            sqlDataDialer,
 		userAgent:                strings.Join(cfg.useragents, " "),
 		applicationName:          cfg.applicationName,
 	}
@@ -418,7 +448,7 @@ func (d *Dialer) Dial(ctx context.Context, icn string, opts ...DialOption) (conn
 	attrs := tel.Attributes{
 		IAMAuthN:    cfg.useIAMAuthN,
 		RefreshType: tel.RefreshAheadType,
-		IPType:      cfg.ipType,
+		IPType:      cfg.connectionType,
 	}
 	if d.lazyRefresh {
 		attrs.RefreshType = tel.RefreshLazyType
@@ -441,6 +471,65 @@ func (d *Dialer) Dial(ctx context.Context, icn string, opts ...DialOption) (conn
 		d.logger.Debugf(ctx, "resolved instance connection string %s to %s", icn, cn.String())
 	}
 
+	// When using NoIP connections, there is no need to retrieve connection
+	// info and instead we may proceed immediately to connecting.
+	if cfg.connectionType == cloudsql.SQLData {
+		key := createKey(cn)
+		d.lock.RLock()
+		allowed, connInCache := d.sqlDataConnAllowed[key]
+		d.lock.RUnlock()
+		// if the cache an entry, and the cache entry says sqlDataService is not allowed
+		if connInCache && !allowed {
+			// fall back to AutoIP
+			cfg.connectionType = cloudsql.AutoIP
+		} else {
+			// Attempt to connect SqlDataService
+			sdcConn, sdcErr := d.connectSQLDataService(ctx, cn, cfg, startTime, mr, attrs)
+			if sdcErr == nil {
+				// Connection succeeded, return the connection.
+				return sdcConn, nil
+			}
+			// Connection failed.
+
+			if !isPreconditionFailed(sdcErr) {
+				return nil, sdcErr
+			}
+			// sdcErr is a PreconditionFailed error. Fall back to auto-ip
+			// This is a streaming gRPC. The PreconditionFailed error usually occurs on the
+			// first read from the stream, not here.
+			cfg.connectionType = cloudsql.AutoIP
+		}
+	}
+
+	return d.connectInstanceIP(ctx, cn, cfg, startTime, mr, attrs)
+}
+
+// connectSQLDataService dials a connection through the SqlDataService.
+func (d *Dialer) connectSQLDataService(ctx context.Context, cn instance.ConnName, cfg dialConfig, startTime time.Time, mr tel.MetricRecorder, attrs tel.Attributes) (net.Conn, error) {
+	conn, err := d.sqlDataDialer.ConnectSQLDataService(ctx, cn)
+	if err != nil {
+		return nil, err
+	}
+
+	// This creates a fallback connection using AutoIP, and marks the connection
+	// as permanently needing to fallback to AutoIP
+	fb := func() (net.Conn, error) {
+		key := createKey(cn)
+		d.lock.Lock()
+		d.sqlDataConnAllowed[key] = false
+		d.lock.Unlock()
+		cfg.connectionType = cloudsql.AutoIP
+		return d.connectInstanceIP(ctx, cn, cfg, startTime, mr, attrs)
+	}
+
+	// Assuming the connection is successful if Recv doesn't return an error
+	// immediately. The first message might not contain data, so we don't pass
+	// any initialData.
+	return newFallbackConn(conn, isSQLDataUnsupportedError, fb), nil
+}
+
+// connectInstanceIP dials the IP address of the instance using a normal net.Conn.
+func (d *Dialer) connectInstanceIP(ctx context.Context, cn instance.ConnName, cfg dialConfig, startTime time.Time, mr tel.MetricRecorder, attrs tel.Attributes) (conn net.Conn, err error) {
 	var endInfo trace.EndSpanFunc
 	ctx, endInfo = trace.StartSpan(ctx, "cloud.google.com/go/cloudsqlconn/internal.InstanceInfo")
 	c, cacheHit, err := d.connectionInfoCache(ctx, cn, &cfg.useIAMAuthN)
@@ -476,11 +565,34 @@ func (d *Dialer) Dial(ctx context.Context, icn string, opts ...DialOption) (conn
 	var connectEnd trace.EndSpanFunc
 	ctx, connectEnd = trace.StartSpan(ctx, "cloud.google.com/go/cloudsqlconn/internal.Connect")
 	defer func() { connectEnd(err) }()
-	addr, err := ci.Addr(cfg.ipType)
+
+	addr, err := ci.Addr(cfg.connectionType)
 	if err != nil {
 		d.removeCached(ctx, cn, c, err)
 		return nil, err
 	}
+
+	// If the connector is configured with a custom DNS name, attempt to use
+	// that DNS name to connect to the instance. Fall back to the metadata IP
+	// address if the DNS name does not resolve to an IP address.
+	if cn.HasDomainName() {
+		addrs, err := d.dnsResolver.LookupHost(ctx, cn.DomainName())
+		if err != nil {
+			d.logger.Debugf(ctx,
+				"[%v] custom DNS name %q did not resolve to an IP address: %v, using %s from instance metadata",
+				cn.String(), cn.DomainName(), err, addr)
+		} else if len(addrs) == 0 {
+			d.logger.Debugf(ctx,
+				"[%v] custom DNS name %q resolved but returned no entries, using %s from instance metadata",
+				cn.String(), cn.DomainName(), addr)
+		} else {
+			d.logger.Debugf(ctx,
+				"[%v] custom DNS name %q resolved to %q, using it to connect",
+				cn.String(), cn.DomainName(), addrs[0])
+			addr = addrs[0]
+		}
+	}
+
 	addr = net.JoinHostPort(addr, serverProxyPort)
 	f := d.dialFunc
 	if cfg.dialFunc != nil {
@@ -526,7 +638,7 @@ func (d *Dialer) Dial(ctx context.Context, icn string, opts ...DialOption) (conn
 	latency := time.Since(startTime).Milliseconds()
 	n := c.openConnsCount.Add(1)
 	trace.RecordOpenConnections(ctx, int64(n), d.dialerID, cn.String())
-	trace.RecordDialLatency(ctx, icn, d.dialerID, latency)
+	trace.RecordDialLatency(ctx, cn.String(), d.dialerID, latency)
 	mr.RecordOpenConnection(ctx, attrs)
 	mr.RecordConnectLatencies(ctx, attrs, latency)
 
@@ -565,6 +677,15 @@ func (d *Dialer) Dial(ctx context.Context, icn string, opts ...DialOption) (conn
 	d.logger.Debugf(ctx, "dial successful")
 	return iConn, nil
 }
+
+func isSQLDataUnsupportedError(err error) bool {
+	if s, ok := status.FromError(err); ok {
+		return s.Code() == codes.FailedPrecondition &&
+			strings.HasPrefix(s.Message(), "unsupported instance edition:")
+	}
+	return false
+}
+
 func (d *Dialer) isTLSError(err error) bool {
 	if nErr, ok := err.(net.Error); ok {
 		return !nErr.Timeout() && // it's a permanent net error
@@ -780,6 +901,10 @@ func (d *Dialer) Close() error {
 	for _, i := range d.cache {
 		i.Close()
 	}
+	if err := d.sqlDataDialer.Close(); err != nil {
+		d.logger.Debugf(context.Background(),
+			"failed to close SQL data connection: %v", err)
+	}
 	return nil
 }
 
@@ -892,5 +1017,12 @@ func newMDXRequest(ci cloudsql.ConnectionInfo, cfg dialConfig, metadataExchangeD
 		return nil
 	}
 
-	return &mdx.MetadataExchangeRequest{ClientProtocolType: &cpt}
+	return mdx.MetadataExchangeRequest_builder{ClientProtocolType: &cpt}.Build()
+}
+
+func isPreconditionFailed(err error) bool {
+	if s, ok := status.FromError(err); ok {
+		return s.Code() == codes.FailedPrecondition
+	}
+	return false
 }